Env refactor
