+++ package.json
@@ -23,7 +23,6 @@
   "dependencies": {
     "@astrojs/tailwind": "^6.0.2",
     "@tailwindcss/vite": "^4.1.11",
-    "@vitest/coverage-v8": "^3.2.4",
     "astro": "^5.11.1",
     "diff": "^8.0.2",
     "diff-match-patch": "^1.0.5",
@@ -33,6 +32,7 @@
     "winston": "^3.17.0"
   },
   "devDependencies": {
+    "@vitest/coverage-v8": "^3.2.4",
     "@testing-library/jest-dom": "^6.6.4",
     "@testing-library/react": "^16.3.0",
     "@types/diff": "^5.2.1",
