+++ src/lib/github.ts
@@ -1,26 +1,6 @@
 import { z } from 'astro/zod';
 import { logger } from './logger';
-
-// Handle import.meta.env gracefully for Jest environment
-const getEnvVar = (key: string, fallback?: string) => {
-  // Check if we're in a test environment first
-  if (process.env.NODE_ENV === 'test') {
-    return process.env[key] || fallback;
-  }
-  
-  // In non-test environments, use import.meta.env if available
-  try {
-    // Use dynamic access to avoid Jest parse errors
-    const importMetaEnv = (globalThis as any).import?.meta?.env || (global as any).import?.meta?.env;
-    if (importMetaEnv) {
-      return importMetaEnv[key] || fallback;
-    }
-  } catch (e) {
-    // Fallback to process.env if import.meta is not available
-  }
-  
-  return process.env[key] || fallback;
-};
+import { getEnvVar } from './env';
 
 const GITHUB_TOKEN = getEnvVar('GITHUB_TOKEN');
 const REPO_OWNER = getEnvVar('GITHUB_REPO_OWNER', 'bgreenawald');
