+++ src/types/astro.d.ts
@@ -2,4 +2,4 @@ declare module '*.astro' {
   import type { ComponentInstance } from 'astro';
   const Component: ComponentInstance;
   export default Component;
-} 
+    }
